use long long

+++ ios/AmplitudeSDK.m
@@ -39,7 +39,7 @@ RCT_REMAP_METHOD(getDeviceId, getDeviceIdWithResolver:(RCTPromiseResolveBlock)re
 
 RCT_REMAP_METHOD(getSessionId, getSessionIdWithResolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
 {
-    long sessionId = [[Amplitude instance] getSessionId];
+    long long sessionId = [[Amplitude instance] getSessionId];
     resolve(sessionId);
 }
